Moved some utility functions into util.py.
I don't think they belong with the other node-related code in node.py.
