+++ src/datahandlers/ensembl.py
@@ -1,6 +1,7 @@
 import traceback
 
-from src.babel_utils import make_local_name, get_config
+from src.babel_utils import make_local_name
+from src.util import get_config
 from apybiomart import find_datasets, query, find_attributes
 import os
