+++ src/exporters/duckdb_exporters.py
@@ -9,7 +9,7 @@ from collections import Counter, defaultdict
 
 import duckdb
 
-from src.node import get_config
+from src.util import get_config
 
 
 def setup_duckdb(duckdb_filename):
